requirements.txt
Flask==1.1.1
flask-restplus==0.13.0
behave==1.2.6
PyHamcrest==1.9.0
allure-behave==2.8.13
Werkzeug==0.16.1
gunicorn==20.0.2
requests==2.23.0